Fix react warnings
Two flavours of warning:

1) The classic of having to use "className" instead of "class" because it's the
DOM property not the HTML attribute name

2) React is fussy about lists of
children if the children don't have keys. This isn't a problem for us because
we're rendering everything statically, so it doesn't matter if react can't track
which element is which, but it still complains. Adding React.Children.toArray fixes
this.