Compile translations YAML to typescript definition
This allows us to use the translations YAML and get full type support for it
without needing to hand roll the typescript definition.

The compiler is not perfect (e.g. it finds two things that it wants to call
Navigationlink, so the one we actually want gets called "Navigationlink2"), but
it's ~okay.